requirements.txt
fastapi
uvicorn
ply
python-dotenv
azure-cosmos
pytest
pydantic
starlette
typing-extensions
python-multipart
google-generativeai
